Improve event tests
